+++ Modules/AddOn/app/Http/Controllers/AddOnController.php
@@ -4,15 +4,17 @@ namespace Modules\AddOn\Http\Controllers;
 
 use App\Http\Controllers\Controller;
 use App\Traits\ApiResponse;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Modules\AddOn\Http\Requests\CreateAddOnRequest;
 use Modules\AddOn\Http\Requests\UpdateAddOnRequest;
 use Modules\AddOn\Http\Resources\AddOnResource;
 use Modules\AddOn\Services\AddOnService;
+use Modules\AddOn\Models\AddOn;
 use Illuminate\Http\JsonResponse;
 
 class AddOnController extends Controller
 {
-    use ApiResponse;
+    use ApiResponse, AuthorizesRequests;
 
     public function __construct(protected AddOnService $addOnService)
     {
@@ -23,6 +25,7 @@ class AddOnController extends Controller
      */
     public function index(): JsonResponse
     {
+        $this->authorize('viewAny', AddOn::class);
         $addOns = $this->addOnService->getAllAddOns();
 
         return $this->successResponse([
@@ -35,6 +38,7 @@ class AddOnController extends Controller
      */
     public function store(CreateAddOnRequest $request): JsonResponse
     {
+        $this->authorize('create', AddOn::class);
         $addOn = $this->addOnService->createAddOn($request->all());
 
         return $this->successResponse([
@@ -48,6 +52,7 @@ class AddOnController extends Controller
     public function show(int $id): JsonResponse
     {
         $addOn = $this->addOnService->getAddOnById($id);
+        $this->authorize('view', $addOn);
 
         return $this->successResponse([
             "add_on" => new AddOnResource($addOn),
@@ -59,6 +64,8 @@ class AddOnController extends Controller
      */
     public function update(UpdateAddOnRequest $request, int $id): JsonResponse
     {
+        $addOn = $this->addOnService->getAddOnById($id);
+        $this->authorize('update', $addOn);
         $addOn = $this->addOnService->updateAddOn($id, $request->all());
 
         return $this->successResponse([
@@ -71,6 +78,8 @@ class AddOnController extends Controller
      */
     public function destroy(int $id): JsonResponse
     {
+        $addOn = $this->addOnService->getAddOnById($id);
+        $this->authorize('delete', $addOn);
         $this->addOnService->deleteAddOn($id);
 
         return $this->successResponse(null, __("message.success"));
