+++ Modules/Category/app/Http/Controllers/Admin/CategoryController.php
@@ -4,16 +4,18 @@ namespace Modules\Category\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Traits\ApiResponse;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Modules\Category\Http\Requests\CreateCategoryRequest;
 use Modules\Category\Http\Requests\UpdateCategoryRequest;
 use Modules\Category\Http\Resources\CategoryResource;
 use Modules\Category\Services\CategoryService;
+use Modules\Category\Models\Category;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 
 class CategoryController extends Controller
 {
-    use ApiResponse;
+    use ApiResponse, AuthorizesRequests;
     public function __construct(protected CategoryService $categoryService)
     {
     }
@@ -23,6 +25,7 @@ class CategoryController extends Controller
      */
     public function index(Request $request): JsonResponse
     {
+        $this->authorize('viewAny', Category::class);
         $filters = $request->all();
         $categories = $this->categoryService->getAllCountries($filters);
         return $this->successResponse([
@@ -35,6 +38,7 @@ class CategoryController extends Controller
      */
     public function store(CreateCategoryRequest $request): JsonResponse
     {
+        $this->authorize('create', Category::class);
         $category = $this->categoryService->createCategory($request->all());
         return $this->successResponse([
             "category"=> new CategoryResource($category),
@@ -47,6 +51,7 @@ class CategoryController extends Controller
     public function show(int $id): JsonResponse
     {
         $category = $this->categoryService->getCategoryById($id);
+        $this->authorize('view', $category);
         return $this->successResponse([
             "category"=> new CategoryResource($category),
             // 'subcategories' => CategoryResource::collection($category->children),
@@ -58,6 +63,8 @@ class CategoryController extends Controller
      */
     public function update(UpdateCategoryRequest $request, int $id): JsonResponse
     {
+        $category = $this->categoryService->getCategoryById($id);
+        $this->authorize('update', $category);
         $category = $this->categoryService->updateCategory($id, $request->all());
         return $this->successResponse([
             "category"=> new CategoryResource($category),
@@ -69,6 +76,8 @@ class CategoryController extends Controller
      */
     public function destroy(int $id): JsonResponse
     {
+        $category = $this->categoryService->getCategoryById($id);
+        $this->authorize('delete', $category);
         $deleted = $this->categoryService->deleteCategory($id);
         return $this->successResponse(null, __("message.success"));
     }
