+++ Modules/City/app/Http/Controllers/Admin/CityController.php
@@ -3,16 +3,18 @@
 namespace Modules\City\Http\Controllers\Admin;
 
 use App\Traits\ApiResponse;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Modules\City\Services\CityService;
 use Modules\City\Http\Resources\CityResource;
 use Modules\City\Http\Requests\CreateCityRequest;
 use Modules\City\Http\Requests\UpdateCityRequest;
+use Modules\City\Models\City;
 
 class CityController extends Controller
 {
-    use ApiResponse;
+    use ApiResponse, AuthorizesRequests;
 
     // Injecting CityService to manage business logic
     public function __construct(private CityService $cityService) {}
@@ -22,6 +24,7 @@ class CityController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', City::class);
         $cities = $this->cityService->getAllCitys(request()->all());
         return $this->successResponse([
             "cities" => CityResource::collection($cities),
@@ -33,6 +36,7 @@ class CityController extends Controller
      */
     public function store(CreateCityRequest $request)
     {
+        $this->authorize('create', City::class);
         $city = $this->cityService->createCity($request->all());
         return $this->successResponse([
             "city" => new CityResource($city),
@@ -45,6 +49,7 @@ class CityController extends Controller
     public function show($id)
     {
         $city = $this->cityService->getCityById($id);
+        $this->authorize('view', $city);
         return $this->successResponse([
             "city" => new CityResource($city->load('governorate')),
         ], __("message.success"));
@@ -55,6 +60,8 @@ class CityController extends Controller
      */
     public function update(UpdateCityRequest $request, $id)
     {
+        $city = $this->cityService->getCityById($id);
+        $this->authorize('update', $city);
         $city = $this->cityService->updateCity($id, $request->all());
         return $this->successResponse([
             "city" => new CityResource($city->load('governorate')),
@@ -66,6 +73,8 @@ class CityController extends Controller
      */
     public function destroy($id)
     {
+        $city = $this->cityService->getCityById($id);
+        $this->authorize('delete', $city);
         $deleted = $this->cityService->deleteCity($id);
         return $this->successResponse(null, __("message.success"));
     }
