+++ Modules/Country/app/Http/Controllers/Admin/CountryController.php
@@ -3,6 +3,7 @@
 namespace Modules\Country\Http\Controllers\Admin;
 
 use App\Traits\ApiResponse;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\PaginationResource;
 use Illuminate\Http\Request;
@@ -10,10 +11,11 @@ use Modules\Country\Services\CountryService;
 use Modules\Country\Http\Resources\CountryResource;
 use Modules\Country\Http\Requests\StoreCountryRequest;
 use Modules\Country\Http\Requests\UpdateCountryRequest;
+use Modules\Country\Models\Country;
 
 class CountryController extends Controller
 {
-    use ApiResponse;
+    use ApiResponse, AuthorizesRequests;
     public function __construct(private CountryService $countryService) {}
 
     /**
@@ -21,6 +23,7 @@ class CountryController extends Controller
      */
     public function index(Request $request)
     {
+        $this->authorize('viewAny', Country::class);
         $filters = $request->only(['search']);
         $countries = $this->countryService->getAllCountries($filters);
         return $this->successResponse([
@@ -33,6 +36,7 @@ class CountryController extends Controller
      */
     public function store(StoreCountryRequest $request)
     {
+        $this->authorize('create', Country::class);
         $country = $this->countryService->createCountry($request->validated());
         return $this->successResponse([
             "country" => new CountryResource($country),
@@ -45,6 +49,7 @@ class CountryController extends Controller
     public function show($id)
     {
         $country = $this->countryService->getCountryById($id);
+        $this->authorize('view', $country);
         return $this->successResponse([
             "country"=> new CountryResource($country),
         ], __("message.success"));
@@ -55,6 +60,8 @@ class CountryController extends Controller
      */
     public function update(UpdateCountryRequest $request, $id)
     {
+        $country = $this->countryService->getCountryById($id);
+        $this->authorize('update', $country);
         $country = $this->countryService->updateCountry($id, $request->all());
         return $this->successResponse([
             "country"=> new CountryResource($country),
@@ -66,6 +73,8 @@ class CountryController extends Controller
      */
     public function destroy($id)
     {
+        $country = $this->countryService->getCountryById($id);
+        $this->authorize('delete', $country);
         $deleted = $this->countryService->deleteCountry($id);
         return $this->successResponse(null, __("message.success"));
     }
