+++ Modules/Coupon/app/Http/Controllers/Admin/CouponController.php
@@ -5,16 +5,18 @@ namespace Modules\Coupon\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\PaginationResource;
 use App\Traits\ApiResponse;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Modules\Coupon\Http\Requests\CreateCouponRequest;
 use Modules\Coupon\Http\Requests\UpdateCouponRequest;
 use Modules\Coupon\Http\Resources\CouponResource;
 use Modules\Coupon\Services\CouponService;
+use Modules\Coupon\Models\Coupon;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 
 class CouponController extends Controller
 {
-    use ApiResponse;
+    use ApiResponse, AuthorizesRequests;
 
     public function __construct(protected CouponService $couponService)
     {
@@ -25,6 +27,7 @@ class CouponController extends Controller
      */
     public function index(Request $request): JsonResponse
     {
+        $this->authorize('viewAny', Coupon::class);
         $filters = $request->all();
         $coupons = $this->couponService->getAllCoupons($filters);
         return $this->successResponse([
@@ -38,6 +41,7 @@ class CouponController extends Controller
      */
     public function store(CreateCouponRequest $request): JsonResponse
     {
+        $this->authorize('create', Coupon::class);
         $coupon = $this->couponService->createCoupon($request->all());
         return $this->successResponse([
             "coupon" => new CouponResource($coupon),
@@ -50,6 +54,7 @@ class CouponController extends Controller
     public function show(int $id): JsonResponse
     {
         $coupon = $this->couponService->getCouponById($id);
+        $this->authorize('view', $coupon);
         return $this->successResponse([
             "coupon" => new CouponResource($coupon),
         ], __("message.success"));
@@ -60,6 +65,8 @@ class CouponController extends Controller
      */
     public function update(UpdateCouponRequest $request, int $id): JsonResponse
     {
+        $coupon = $this->couponService->getCouponById($id);
+        $this->authorize('update', $coupon);
         $coupon = $this->couponService->updateCoupon($id, $request->all());
         return $this->successResponse([
             "coupon" => new CouponResource($coupon),
@@ -71,6 +78,8 @@ class CouponController extends Controller
      */
     public function destroy(int $id): JsonResponse
     {
+        $coupon = $this->couponService->getCouponById($id);
+        $this->authorize('delete', $coupon);
         $deleted = $this->couponService->deleteCoupon($id);
         return $this->successResponse(null, __("message.success"));
     }
