+++ Modules/Governorate/app/Http/Controllers/GovernorateController.php
@@ -4,15 +4,17 @@ namespace Modules\Governorate\Http\Controllers;
 
 use App\Http\Controllers\Controller;
 use App\Traits\ApiResponse;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 use Modules\Governorate\Services\GovernorateService;
 use Modules\Governorate\Http\Requests\StoreGovernorateRequest;
 use Modules\Governorate\Http\Requests\UpdateGovernorateRequest;
 use Modules\Governorate\Http\Resources\GovernorateResource;
+use Modules\Governorate\Models\Governorate;
 
 class GovernorateController extends Controller
 {
-    use ApiResponse;
+    use ApiResponse, AuthorizesRequests;
 
     // Injecting GovernorateService to manage business logic
     public function __construct(private GovernorateService $governorateService) {}
@@ -22,6 +24,7 @@ class GovernorateController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', Governorate::class);
         $governorates = $this->governorateService->getAllGovernorates(request()->all());
         return $this->successResponse([
             "governorates" => GovernorateResource::collection($governorates->load('country')),
@@ -33,6 +36,7 @@ class GovernorateController extends Controller
      */
     public function store(StoreGovernorateRequest $request)
     {
+        $this->authorize('create', Governorate::class);
         $governorate = $this->governorateService->createGovernorate($request->all());
         return $this->successResponse([
             "governorate" => new GovernorateResource($governorate),
@@ -45,6 +49,7 @@ class GovernorateController extends Controller
     public function show($id)
     {
         $governorate = $this->governorateService->getGovernorateById($id);
+        $this->authorize('view', $governorate);
         return $this->successResponse([
             "governorate"=> new GovernorateResource($governorate->load('country')),
         ], __("message.success"));
@@ -55,6 +60,8 @@ class GovernorateController extends Controller
      */
     public function update(UpdateGovernorateRequest $request, $id)
     {
+        $governorate = $this->governorateService->getGovernorateById($id);
+        $this->authorize('update', $governorate);
         $governorate = $this->governorateService->updateGovernorate($id, $request->all());
         return $this->successResponse([
             "governorate"=> new GovernorateResource($governorate->load('country')),
@@ -66,6 +73,8 @@ class GovernorateController extends Controller
      */
     public function destroy($id)
     {
+        $governorate = $this->governorateService->getGovernorateById($id);
+        $this->authorize('delete', $governorate);
         $deleted = $this->governorateService->deleteGovernorate($id);
         return $this->successResponse(null, __("message.success"));
     }
