+++ Modules/LoyaltySetting/app/Http/Controllers/LoyaltySettingController.php
@@ -4,15 +4,17 @@ namespace Modules\LoyaltySetting\Http\Controllers;
 
 use App\Http\Controllers\Controller;
 use App\Traits\ApiResponse;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 use Modules\LoyaltySetting\Services\LoyaltySettingService;
 use Modules\LoyaltySetting\Http\Requests\StoreLoyaltySettingRequest;
 use Modules\LoyaltySetting\Http\Requests\UpdateLoyaltySettingRequest;
 use Modules\LoyaltySetting\Http\Resources\LoyaltySettingResource;
+use Modules\LoyaltySetting\Models\LoyaltySetting;
 
 class LoyaltySettingController extends Controller
 {
-    use ApiResponse;
+    use ApiResponse, AuthorizesRequests;
 
     // Injecting LoyaltySettingService to manage business logic
     public function __construct(private LoyaltySettingService $LoyaltySettingService) {}
@@ -22,6 +24,7 @@ class LoyaltySettingController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', LoyaltySetting::class);
         $LoyaltySettings = $this->LoyaltySettingService->getAllLoyaltySettings();
         return $this->successResponse([
             "LoyaltySettings" => LoyaltySettingResource::collection($LoyaltySettings->load('country')),
@@ -33,6 +36,7 @@ class LoyaltySettingController extends Controller
      */
     public function store(StoreLoyaltySettingRequest $request)
     {
+        $this->authorize('create', LoyaltySetting::class);
         $LoyaltySetting = $this->LoyaltySettingService->createLoyaltySetting($request->all());
         return $this->successResponse([
             "LoyaltySetting" => new LoyaltySettingResource($LoyaltySetting),
@@ -45,6 +49,7 @@ class LoyaltySettingController extends Controller
     public function show($id)
     {
         $LoyaltySetting = $this->LoyaltySettingService->getLoyaltySettingById($id);
+        $this->authorize('view', $LoyaltySetting);
         return $this->successResponse([
             "LoyaltySetting"=> new LoyaltySettingResource($LoyaltySetting),
         ], __("message.success"));
@@ -55,6 +60,8 @@ class LoyaltySettingController extends Controller
      */
     public function update(UpdateLoyaltySettingRequest $request, $id)
     {
+        $LoyaltySetting = $this->LoyaltySettingService->getLoyaltySettingById($id);
+        $this->authorize('update', $LoyaltySetting);
         $LoyaltySetting = $this->LoyaltySettingService->updateLoyaltySetting($id, $request->all());
         return $this->successResponse([
             "LoyaltySetting"=> new LoyaltySettingResource($LoyaltySetting),
@@ -66,6 +73,8 @@ class LoyaltySettingController extends Controller
      */
     public function destroy($id)
     {
+        $LoyaltySetting = $this->LoyaltySettingService->getLoyaltySettingById($id);
+        $this->authorize('delete', $LoyaltySetting);
         $deleted = $this->LoyaltySettingService->deleteLoyaltySetting($id);
         return $this->successResponse(null, __("message.success"));
     }
