+++ Modules/Option/app/Http/Controllers/Admin/OptionController.php
@@ -3,16 +3,18 @@
 namespace Modules\Option\Http\Controllers\Admin;
 
 use App\Traits\ApiResponse;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Modules\Option\Services\OptionService;
 use Modules\Option\Http\Resources\OptionResource;
 use Modules\Option\Http\Requests\CreateOptionRequest;
 use Modules\Option\Http\Requests\UpdateOptionRequest;
+use Modules\Option\Models\Option;
 
 class OptionController extends Controller
 {
-    use ApiResponse;
+    use ApiResponse, AuthorizesRequests;
 
     public function __construct(private OptionService $optionService) {}
 
@@ -21,6 +23,7 @@ class OptionController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', Option::class);
         $options = $this->optionService->getAllOptions();
         return $this->successResponse([
             'options' => OptionResource::collection($options),
@@ -32,6 +35,7 @@ class OptionController extends Controller
      */
     public function store(CreateOptionRequest $request)
     {
+        $this->authorize('create', Option::class);
         $option = $this->optionService->createOption($request->validated());
         return $this->successResponse([
             'option' => new OptionResource($option),
@@ -44,6 +48,7 @@ class OptionController extends Controller
     public function show($id)
     {
         $option = $this->optionService->getOptionById($id);
+        $this->authorize('view', $option);
         return $this->successResponse([
             'option' => new OptionResource($option),
         ], __('message.success'));
@@ -54,6 +59,8 @@ class OptionController extends Controller
      */
     public function update(UpdateOptionRequest $request, $id)
     {
+        $option = $this->optionService->getOptionById($id);
+        $this->authorize('update', $option);
         $option = $this->optionService->updateOption($id, $request->all());
         return $this->successResponse([
             'option' => new OptionResource($option),
@@ -65,6 +72,8 @@ class OptionController extends Controller
      */
     public function destroy($id)
     {
+        $option = $this->optionService->getOptionById($id);
+        $this->authorize('delete', $option);
         $this->optionService->deleteOption($id);
         return $this->successResponse(null, __('message.success'));
     }
